voicecode-backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
websockets==12.0
httpx==0.26.0

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.0
firebase-admin==6.4.0
google-cloud-aiplatform==1.42.0
google-generativeai==0.3.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

